feat: create filters hook
